getBuiltDockerCompose function
